Mensaje de confirmación antes de eliminar un archivo o versión

+++ src/main/java/es/jklabs/gui/menu/contextual/S3FilePopUp.java
@@ -14,6 +14,8 @@ import java.util.Objects;
 public class S3FilePopUp extends JPopupMenu {
 
     private static final long serialVersionUID = 5925867780383236170L;
+    private static final String ELIMINAR = "eliminar";
+    private static final String IMG_ICONS_TRASH_PNG = "img/icons/trash.png";
     private final Explorador explorador;
     private final S3File s3File;
 
@@ -30,7 +32,7 @@ public class S3FilePopUp extends JPopupMenu {
         JMenuItem jmiDescargar = new JMenuItem(Mensajes.getMensaje("descargar"), iconDownload);
         jmiDescargar.addActionListener(l -> descargarArchivo());
         add(jmiDescargar);
-        ImageIcon iconPapelera = getIcon("img/icons/trash.png");
+        ImageIcon iconPapelera = getIcon(IMG_ICONS_TRASH_PNG);
         ImageIcon iconReloj = getIcon("img/icons/clock.png");
         JMenu jmVersiones = new JMenu(Mensajes.getMensaje("versiones"));
         jmVersiones.setIcon(iconReloj);
@@ -38,7 +40,7 @@ public class S3FilePopUp extends JPopupMenu {
         for (S3FileVersion s3FileVersion : s3FileVersionList) {
             JMenu jmRegVersion = new JMenu(s3FileVersion.getFecha().toString());
             JMenuItem jmiDownloadVersion = new JMenuItem(Mensajes.getMensaje("descargar"), iconDownload);
-            JMenuItem jmiDeleteVersion = new JMenuItem(Mensajes.getMensaje("eliminar"), iconPapelera);
+            JMenuItem jmiDeleteVersion = new JMenuItem(Mensajes.getMensaje(ELIMINAR), iconPapelera);
             jmiDownloadVersion.addActionListener(l -> descargarVersion(s3FileVersion));
             jmiDeleteVersion.addActionListener(l -> eliminarVersion(s3FileVersion));
             jmRegVersion.add(jmiDownloadVersion);
@@ -46,7 +48,7 @@ public class S3FilePopUp extends JPopupMenu {
             jmVersiones.add(jmRegVersion);
         }
         add(jmVersiones);
-        JMenuItem jmiEliminar = new JMenuItem(Mensajes.getMensaje("eliminar"), iconPapelera);
+        JMenuItem jmiEliminar = new JMenuItem(Mensajes.getMensaje(ELIMINAR), iconPapelera);
         jmiEliminar.addActionListener(l -> elminarArchivo());
         add(jmiEliminar);
     }
@@ -63,13 +65,25 @@ public class S3FilePopUp extends JPopupMenu {
     }
 
     private void eliminarVersion(S3FileVersion s3FileVersion) {
-        UtilidadesS3.elimninarVersion(explorador.getPadre().getConfiguracion().getBucketConfig(), s3File, s3FileVersion);
-        explorador.recargarPantalla();
+        ImageIcon icon = new ImageIcon(Objects.requireNonNull(getClass().getClassLoader().getResource
+                (IMG_ICONS_TRASH_PNG)));
+        int input = JOptionPane.showConfirmDialog(explorador, Mensajes.getMensaje("confirmacion.eliminar", new String[]{s3FileVersion.getFecha().toString()}), Mensajes.getMensaje(ELIMINAR),
+                JOptionPane.YES_NO_OPTION, JOptionPane.QUESTION_MESSAGE, icon);
+        if (input == JOptionPane.YES_OPTION) {
+            UtilidadesS3.elimninarVersion(explorador.getPadre().getConfiguracion().getBucketConfig(), s3File, s3FileVersion);
+            explorador.recargarPantalla();
+        }
     }
 
     private void elminarArchivo() {
-        UtilidadesS3.deleteObject(explorador.getPadre().getConfiguracion().getBucketConfig(), s3File);
-        explorador.recargarPantalla();
+        ImageIcon icon = new ImageIcon(Objects.requireNonNull(getClass().getClassLoader().getResource
+                (IMG_ICONS_TRASH_PNG)));
+        int input = JOptionPane.showConfirmDialog(explorador, Mensajes.getMensaje("confirmacion.eliminar", new String[]{s3File.getName()}), Mensajes.getMensaje(ELIMINAR),
+                JOptionPane.YES_NO_OPTION, JOptionPane.QUESTION_MESSAGE, icon);
+        if (input == JOptionPane.YES_OPTION) {
+            UtilidadesS3.deleteObject(explorador.getPadre().getConfiguracion().getBucketConfig(), s3File);
+            explorador.recargarPantalla();
+        }
     }
 
     private void descargarArchivo() {
